(#17716) Changed to HPUX user provider to handle password changes
The user provider hpux.rb did not have the ability to manage
passwords so that on each check by puppet the password was set to
:absent by the pw.rb password provider script as it tried to
check the /etc/shadow file.  HPUX trusted computing does not use
the /etc/shadow files. As a result, even when the password hash
was correct it still changed the passwords for every user.

The updates add in the :manages_passwords field to the
has_features and created logic to be able to get the password
has from either the /etc/password when trusted computing is off
or get the password had from the /tcb/files directories when
trusted computing is turned on.

The hpux_spec.rb file was updated to check that manage_passwords
is set, the user exists and the provider returns a password.
To ensure a correct result on the password return section the
root user is specified as it should always exist on an HPUX
system.